+++ Runtime/Scripts/Settings/SelectionGroupsEditorProjectSettings.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections.Generic;
 using Unity.FilmInternalUtilities;
 using UnityEngine;
+using UnityEngine.Assertions;
 
 
 namespace Unity.SelectionGroups {
@@ -13,6 +15,18 @@ internal class SelectionGroupsEditorProjectSettings : BaseJsonSingleton<Selectio
 
     protected override void UpgradeToLatestVersionV(int prevVersion, int curVersion) {
         
+        if (prevVersion < (int) SGProjectSettingsVersion.EDITOR_STATE_0_7_2) {
+#pragma warning disable 612 //obsolete
+            if (null != m_defaultGroupEditorToolStatus) {
+                int numStates = m_defaultGroupEditorToolStatus.Count;
+                for (int i = 0; i < numStates; ++i) {
+                    m_defaultGroupEditorToolStates[i] = m_defaultGroupEditorToolStatus[i];
+                }
+                m_defaultGroupEditorToolStatus = null;
+#pragma warning restore 612
+            }
+        }
+        
     }
     
 //----------------------------------------------------------------------------------------------------------------------
@@ -23,14 +37,32 @@ internal class SelectionGroupsEditorProjectSettings : BaseJsonSingleton<Selectio
         m_groupsVisibleInHierarchy = visible; 
     }
 
+    internal bool GetDefaultGroupEditorToolState(int toolID) {
+        if (m_defaultGroupEditorToolStates.TryGetValue(toolID, out bool status))
+            return status;
+        return false;
+    }
+
+    public void EnableDefaultGroupEditorTool(int toolID, bool toolEnabled) {
+        m_defaultGroupEditorToolStates[toolID] = toolEnabled;
+    }
+    
+    
 //----------------------------------------------------------------------------------------------------------------------
     
-    [SerializeField] private bool m_groupsVisibleInHierarchy = true;
-
+    [SerializeField] private bool m_groupsVisibleInHierarchy       = true;
+    
+    [Obsolete]
+    [SerializeField] List<bool> m_defaultGroupEditorToolStatus = null;
+    
+    [SerializeField] EditorToolStates m_defaultGroupEditorToolStates = new EditorToolStates(); 
+    
 //----------------------------------------------------------------------------------------------------------------------
-    private const int LATEST_VERSION = (int) Version.INITIAL; 
-    enum Version {
+    private const int LATEST_VERSION = (int) SGProjectSettingsVersion.EDITOR_STATE_0_7_2; 
+    enum SGProjectSettingsVersion {
         INITIAL = 1,
+        EDITOR_STATE_0_7_2, //The data structure of m_defaultGroupEditorToolStates was changed
+        
     };