+++ Tests/Runtime/GoQLDiscriminatorTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using NUnit.Framework;
+using Unity.GoQL;
 using UnityEditor.SceneManagement;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -24,6 +25,12 @@ internal class GoQLDiscriminatorTests
     }
     
     [Test]
+    public void ExclusionOperator() {
+        TestUtility.ExecuteGoQLAndVerify("Head<t:Collider><!t:MeshFilter>", 2,
+            (Transform t) => t.name == "Head" && t.GetComponent<Collider>() != null && t.GetComponent<MeshFilter>() == null);
+    }
+    
+    [Test]
     public void GlowMaterial() {
         TestUtility.ExecuteGoQLAndVerify("Head<m:Glow>", 1, (Transform t) => {
             MeshRenderer mr = t.GetComponent<MeshRenderer>();
@@ -37,6 +44,12 @@ internal class GoQLDiscriminatorTests
             MeshRenderer mr = t.GetComponent<MeshRenderer>();
             return t.name == "Head" && null!=mr && mr.sharedMaterial.shader.name == "Standard";
         });
+    }
+    
+    [Test]
+    public void BadDiscriminator() {
+        GoQLExecutor e = new GoQLExecutor("/<Collider/*");
+        Assert.AreEqual(ParseResult.UnexpectedEndOfInput, e.parseResult);
     }
     
 //----------------------------------------------------------------------------------------------------------------------