+++ Tests/Runtime/Scripts/SelectionGroupTests.cs
@@ -1,5 +1,8 @@
+using System.Collections;
 using NUnit.Framework;
+using UnityEditor;
 using UnityEngine;
+using UnityEngine.TestTools;
 
 namespace Unity.SelectionGroups.Tests 
 {
@@ -8,7 +11,7 @@ internal class SelectionGroupTests {
     [Test]
     public void CreateEmptyGroup() {
         SelectionGroupManager groupManager = GetAndInitGroupManager();
-        SelectionGroup        group        = groupManager.CreateSceneSelectionGroup("TestGroup", Color.green);
+        SelectionGroup        group        = groupManager.CreateSelectionGroup("TestGroup", Color.green);
         Assert.IsNotNull(group);
         Assert.AreEqual(1, groupManager.Groups.Count);
     }
@@ -18,7 +21,7 @@ internal class SelectionGroupTests {
     [Test]
     public void DeleteGroupComponent() {
         SelectionGroupManager groupManager = GetAndInitGroupManager();
-        SelectionGroup        group        = groupManager.CreateSceneSelectionGroup("TestGroup", Color.green);       
+        SelectionGroup        group        = groupManager.CreateSelectionGroup("TestGroup", Color.green);       
         Object.DestroyImmediate(group);
         Assert.AreEqual(0, groupManager.Groups.Count);
     }
@@ -26,7 +29,7 @@ internal class SelectionGroupTests {
     [Test]
     public void DeleteGroupGameObject() {
         SelectionGroupManager groupManager = GetAndInitGroupManager();
-        SelectionGroup        group        = groupManager.CreateSceneSelectionGroup("TestGroup", Color.green);       
+        SelectionGroup        group        = groupManager.CreateSelectionGroup("TestGroup", Color.green);       
         Object.DestroyImmediate(group);
         Assert.AreEqual(0, groupManager.Groups.Count);
     }
@@ -34,10 +37,12 @@ internal class SelectionGroupTests {
     [Test]
     public void DeleteGroupByAPI() {
         SelectionGroupManager groupManager = GetAndInitGroupManager();
-        SelectionGroup        group        = groupManager.CreateSceneSelectionGroup("TestGroup", Color.green);
+        SelectionGroup        group        = groupManager.CreateSelectionGroup("TestGroup", Color.green);
         groupManager.DeleteGroup(group);
         Assert.AreEqual(0, groupManager.Groups.Count);
     }
+    
+  
     
     
 //----------------------------------------------------------------------------------------------------------------------